fix: improve some install selection bugs
