+++ src/managers/common/pickers.ts
@@ -2,6 +2,7 @@ import { QuickInputButtons, QuickPickItem, QuickPickItemButtonEvent, QuickPickIt
 import { Common, PackageManagement } from '../../common/localize';
 import { launchBrowser } from '../../common/env.apis';
 import { showInputBoxWithButtons, showQuickPickWithButtons, showTextDocument } from '../../common/window.apis';
+import { Installable } from './utils';
 
 const OPEN_BROWSER_BUTTON = {
     iconPath: new ThemeIcon('globe'),
@@ -17,50 +18,6 @@ const EDIT_ARGUMENTS_BUTTON = {
     iconPath: new ThemeIcon('pencil'),
     tooltip: PackageManagement.editArguments,
 };
-
-export interface Installable {
-    /**
-     * The name of the package, requirements, lock files, or step name.
-     */
-    readonly name: string;
-
-    /**
-     * The name of the package, requirements, pyproject.toml or any other project file, etc.
-     */
-    readonly displayName: string;
-
-    /**
-     * Arguments passed to the package manager to install the package.
-     *
-     * @example
-     *  ['debugpy==1.8.7'] for `pip install debugpy==1.8.7`.
-     *  ['--pre', 'debugpy'] for `pip install --pre debugpy`.
-     *  ['-r', 'requirements.txt'] for `pip install -r requirements.txt`.
-     */
-    readonly args?: string[];
-
-    /**
-     * Installable group name, this will be used to group installable items in the UI.
-     *
-     * @example
-     *  `Requirements` for any requirements file.
-     *  `Packages` for any package.
-     */
-    readonly group?: string;
-
-    /**
-     * Description about the installable item. This can also be path to the requirements,
-     * version of the package, or any other project file path.
-     */
-    readonly description?: string;
-
-    /**
-     * External Uri to the package on pypi or docs.
-     * @example
-     *  https://pypi.org/project/debugpy/ for `debugpy`.
-     */
-    readonly uri?: Uri;
-}
 
 function handleItemButton(uri?: Uri) {
     if (uri) {
@@ -118,7 +75,7 @@ async function enterPackageManually(filler?: string): Promise<string[] | undefin
 }
 
 interface GroupingResult {
-    items: QuickPickItem[];
+    items: PackageQuickPickItem[];
     installedItems: PackageQuickPickItem[];
 }
 
@@ -132,11 +89,13 @@ function groupByInstalled(items: PackageQuickPickItem[], installed?: string[]): 
             result.push(i);
         }
     });
-    const installedSeparator: QuickPickItem = {
+    const installedSeparator: PackageQuickPickItem = {
+        id: 'installed-sep',
         label: PackageManagement.installed,
         kind: QuickPickItemKind.Separator,
     };
-    const commonPackages: QuickPickItem = {
+    const commonPackages: PackageQuickPickItem = {
+        id: 'common-packages-sep',
         label: PackageManagement.commonPackages,
         kind: QuickPickItemKind.Separator,
     };
@@ -171,7 +130,7 @@ export async function selectFromCommonPackagesToInstall(
     preSelected?: PackageQuickPickItem[] | undefined,
 ): Promise<CommonPackagesResult | undefined> {
     const { installedItems, items } = groupByInstalled(common.map(installableToQuickPickItem), installed);
-    const preSelectedItems = [...installedItems, ...(preSelected ?? [])];
+    const preSelectedItems = items.filter((i) => (preSelected ?? installedItems).some((s) => s.id === i.id));
     let selected: PackageQuickPickItem | PackageQuickPickItem[] | undefined;
     try {
         selected = await showQuickPickWithButtons(
@@ -208,10 +167,8 @@ export async function selectFromCommonPackagesToInstall(
 
     if (selected && Array.isArray(selected)) {
         if (selected.find((s) => s.label === PackageManagement.enterPackageNames)) {
-            const filler = selected
-                .filter((s) => s.label !== PackageManagement.enterPackageNames)
-                .map((s) => s.id)
-                .join(' ');
+            const filtered = selected.filter((s) => s.label !== PackageManagement.enterPackageNames);
+            const filler = filtered.map((s) => s.id).join(' ');
             try {
                 const selections = await enterPackageManually(filler);
                 if (selections) {
@@ -220,7 +177,7 @@ export async function selectFromCommonPackagesToInstall(
                 return undefined;
             } catch (ex) {
                 if (ex === QuickInputButtons.Back) {
-                    return selectFromCommonPackagesToInstall(common, installed, selected);
+                    return selectFromCommonPackagesToInstall(common, installed, filtered);
                 }
                 return undefined;
             }
